CMLComp examples from GULP and SIESTA
The siesta examples are the reference output
files from the tests (each file_name.xml should
match a sub-directory in the siesta/tests directory).

The gulp examples are the first few examples from the
distributed 'examples' directory with 'output cml file.xml'
added to the input dec. There are more gulp examples
in the distribution.